Commit MicroServicio cliente con CRUD
Crud finalizado funcionando

+++ clientes/src/main/java/com/perfumelandiaspa/clientes/Repository/ClienteRepository.java
@@ -1,8 +1,6 @@
 package com.perfumelandiaspa.clientes.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import com.perfumelandiaspa.clientes.Model.Entity.ClienteEntity;
@@ -10,7 +8,7 @@ import com.perfumelandiaspa.clientes.Model.Entity.ClienteEntity;
 
 
 @Repository
-public interface ClienteRepository extends JpaRepository<ClienteEntity, Long> {
+public interface ClienteRepository extends JpaRepository<ClienteEntity,Integer> {
 
     //Metodo de crearCliente
     boolean existsByRut(String rut);
